Refactor: move `CustomContract` attributes into `Attribute` namespace
