+++ src/Populator/CustomContract/Attribute/Populator.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace Neusta\ConverterBundle\Populator\CustomContract;
+namespace Neusta\ConverterBundle\Populator\CustomContract\Attribute;
 
 #[\Attribute(\Attribute::TARGET_METHOD)]
 final class Populator
