+++ src/Populator/CustomContract/Attribute/Source.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace Neusta\ConverterBundle\Populator\CustomContract;
+namespace Neusta\ConverterBundle\Populator\CustomContract\Attribute;
 
 #[\Attribute(\Attribute::TARGET_PARAMETER)]
 final class Source
